fix: best practice

+++ controllers/UserController.js
@@ -3,20 +3,39 @@ import User from "../models/UserModel.js";
 // GET
 async function getUsers(req, res) {
   try {
-    const response = await User.findAll();
-    res.status(200).json(response);
+    const users = await User.findAll();
+    res.status(200).json({
+      status: "Success",
+      message: "Users Retrieved",
+      data: users,
+    });
   } catch (error) {
-    console.log(error.message);
+    res.status(error.statusCode || 500).json({
+      status: "Error",
+      message: error.message,
+    });
   }
 }
 
 // GET BY ID
 async function getUserById(req, res) {
   try {
-    const response = await User.findOne({ where: { id: req.params.id } });
-    res.status(200).json(response);
+    const user = await User.findOne({ where: { id: req.params.id } });
+    if (!user) {
+      const error = new Error("User tidak ditemukan 😮");
+      error.statusCode = 400;
+      throw error;
+    }
+    res.status(200).json({
+      status: "Success",
+      message: "User Retrieved",
+      data: user,
+    });
   } catch (error) {
-    console.log(error.message);
+    res.status(error.statusCode || 500).json({
+      status: "Error",
+      message: error.message,
+    });
   }
 }
 
@@ -25,30 +44,62 @@ async function createUser(req, res) {
   try {
     const inputResult = req.body;
     await User.create(inputResult);
-    res.status(201).json({ msg: "User Created" });
+    res.status(201).json({
+      status: "Success",
+      message: "User Created",
+    });
   } catch (error) {
-    console.log(error.message);
+    res.status(error.statusCode || 500).json({
+      status: "Error",
+      message: error.message,
+    });
   }
 }
 
 async function updateUser(req, res) {
   try {
+    const ifUserExist = await User.findOne({ where: { id: req.params.id } });
+    if (!ifUserExist) {
+      const error = new Error("User tidak ditemukan 😮");
+      error.statusCode = 400;
+      throw error;
+    }
+
     const inputResult = req.body;
     await User.update(inputResult, {
       where: { id: req.params.id },
     });
-    res.status(200).json({ msg: "User Updated" });
+    res.status(200).json({
+      status: "Success",
+      message: "User Updated",
+    });
   } catch (error) {
-    console.log(error.message);
+    res.status(error.statusCode || 500).json({
+      status: "Error",
+      message: error.message,
+    });
   }
 }
 
 async function deleteUser(req, res) {
   try {
+    const ifUserExist = await User.findOne({ where: { id: req.params.id } });
+    if (!ifUserExist) {
+      const error = new Error("User tidak ditemukan 😮");
+      error.statusCode = 400;
+      throw error;
+    }
+
     await User.destroy({ where: { id: req.params.id } });
-    res.status(200).json({ msg: "User Deleted" });
+    res.status(200).json({
+      status: "Success",
+      message: "User Deleted",
+    });
   } catch (error) {
-    console.log(error.message);
+    res.status(error.statusCode || 500).json({
+      status: "Error",
+      message: error.message,
+    });
   }
 }
